pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "cp2077-adtex"
version = "0.1.0"
description = "Config-driven ad texture extract/finalize pipeline for Cyberpunk 2077"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "cp2077_ad_texture_pipeline contributors" }
]
dependencies = [
  "typer>=0.12.0,<1.0.0",
  "rich>=13.7.0,<15.0.0",
  "pillow>=10.3.0,<12.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
]

[project.scripts]
cp2077-adtex = "cp2077_adtex.cli:app"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
